Implement clean architecture + DDD direcotry structure
The main directory (Primary) represents the business domains and in the secondary directory, it consists of domain, persistence, api folders. Common folder will include shared library classes across the domain.
