+++ src/main/java/com/blind/BlindApplication.java
@@ -1,0 +1,11 @@
+package com.blind;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BlindApplication {
+    public static void main(String... args) {
+        SpringApplication.run(BlindApplication.class, args);
+    }
+}
