+++ src/main/java/com/blind/account/api/AccountController.java
@@ -1,0 +1,7 @@
+package com.blind.account.api;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class AccountController {
+}
