+++ src/main/java/com/blind/account/persistence/repository/AccountRepository.java
@@ -1,0 +1,14 @@
+package com.blind.account.persistence.repository;
+
+import com.blind.account.domain.Entity.Account;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface AccountRepository extends JpaRepository<Account, Integer> {
+    Optional<Account> findByEmail(String email);
+    Optional<Account> findByUsername(String username);
+    Optional<Account> findById(String id);
+}
