--- src/main/java/com/blind/api/bookmark/repository/BookmarkRepository.java
@@ -1,9 +1,0 @@
-package com.blind.api.bookmark.repository;
-
-import com.blind.api.bookmark.domain.Bookmark;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface BookmarkRepository extends JpaRepository<Bookmark, Integer> {
-}
