--- src/main/java/com/blind/api/company/domain/CompanyCategory.java
@@ -1,19 +1,0 @@
-package com.blind.api.company.domain;
-
-import com.blind.common.domain.BaseEntity;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.EntityListeners;
-
-@Entity
-@Getter
-@NoArgsConstructor
-@EntityListeners(AuditingEntityListener.class)
-public class CompanyCategory extends BaseEntity {
-    @Column(nullable = false)
-    private String name;
-}
