--- src/main/java/com/blind/api/company/repository/CompanyRepository.java
@@ -1,12 +1,0 @@
-package com.blind.api.company.repository;
-
-import com.blind.api.company.domain.Company;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface CompanyRepository extends JpaRepository<Company, Integer> {
-    Optional<Company> findById(String id);
-}
