--- src/main/java/com/blind/api/job/domain/JobCategory.java
@@ -1,19 +1,0 @@
-package com.blind.api.job.domain;
-
-import com.blind.common.domain.BaseEntity;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.EntityListeners;
-
-@Entity
-@Getter
-@NoArgsConstructor
-@EntityListeners(AuditingEntityListener.class)
-public class JobCategory extends BaseEntity {
-    @Column(nullable = false)
-    private String name;
-}
