--- src/main/java/com/blind/api/post/repository/PostRepository.java
@@ -1,9 +1,0 @@
-package com.blind.api.post.repository;
-
-import com.blind.api.post.domain.Post;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface PostRepository extends JpaRepository<Post, Integer> {
-}
