--- src/main/java/com/blind/api/user/repository/UserRepository.java
@@ -1,14 +1,0 @@
-package com.blind.api.user.repository;
-
-import com.blind.api.user.domain.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
-    Optional<User> findByEmail(String email);
-    Optional<User> findByUsername(String username);
-    Optional<User> findById(String id);
-}
