+++ src/main/java/com/blind/company/api/CompanyController.java
@@ -1,0 +1,7 @@
+package com.blind.company.api;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class CompanyController {
+}
