+++ src/main/java/com/blind/company/domain/Entity/Company.java
@@ -1,0 +1,20 @@
+package com.blind.company.domain.Entity;
+
+import com.blind.common.domain.BaseEntity;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import javax.persistence.*;
+
+@Entity
+@Getter
+@NoArgsConstructor
+@EntityListeners(AuditingEntityListener.class)
+public class Company extends BaseEntity {
+    @Column(nullable=false)
+    private String name;
+    @JoinColumn(name="company_category_id", referencedColumnName = "id", nullable = false)
+    @ManyToOne
+    private CompanyCategory companyCategory;
+}
