+++ src/main/java/com/blind/company/domain/Entity/JobCategory.java
@@ -1,0 +1,19 @@
+package com.blind.company.domain.Entity;
+
+import com.blind.common.domain.BaseEntity;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.EntityListeners;
+
+@Entity
+@Getter
+@NoArgsConstructor
+@EntityListeners(AuditingEntityListener.class)
+public class JobCategory extends BaseEntity {
+    @Column(nullable = false)
+    private String name;
+}
