+++ src/main/java/com/blind/company/persistence/respository/CompanyRepository.java
@@ -1,0 +1,12 @@
+package com.blind.company.persistence.respository;
+
+import com.blind.company.domain.Entity.Company;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface CompanyRepository extends JpaRepository<Company, Integer> {
+    Optional<Company> findById(String id);
+}
