+++ src/main/java/com/blind/company/persistence/respository/JobCategoryRepository.java
@@ -1,0 +1,12 @@
+package com.blind.company.persistence.respository;
+
+import com.blind.company.domain.Entity.JobCategory;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface JobCategoryRepository extends JpaRepository<JobCategory, Integer> {
+    Optional<JobCategory> findById(String id);
+}
