+++ src/main/java/com/blind/company/persistence/respository/JobPositionRepository.java
@@ -1,0 +1,12 @@
+package com.blind.company.persistence.respository;
+
+import com.blind.company.domain.Entity.JobPosition;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface JobPositionRepository extends JpaRepository<JobPosition, Integer> {
+    Optional<JobPosition> findById(String id);
+}
