+++ src/main/java/com/blind/post/api/CommentController.java
@@ -1,0 +1,7 @@
+package com.blind.post.api;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class CommentController {
+}
