+++ src/main/java/com/blind/post/persistence/repository/BookmarkRepository.java
@@ -1,0 +1,9 @@
+package com.blind.post.persistence.repository;
+
+import com.blind.post.domain.Entity.Bookmark;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface BookmarkRepository extends JpaRepository<Bookmark, Integer> {
+}
